refacotor : 키워드 편지 상세 조회 api 연동refactoring
